pyproject.toml
[project]
name = "intent-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "sentence-transformers>=2.7",
  "torch>=2.2",
  "python-dotenv>=1.0",
  "loguru>=0.7.3",
]

[project.scripts]
intent-classifier = "main:main"
intent-worker = "worker:main"

[tool.setuptools]
py-modules = ["main", "clickhouse_check", "worker"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true
